add an option to configure hot-reload (#358)
### 🔦 Summary
Some devs would like to disable `hmr` and also disable `hot-reload`.

In this PR we'll enable it through yoshi config.

### 🗺️ Test Plan
Verify that the code that handles `hot-reload` is disable when the option is false. (in the bundle)

When `hmr` and `hot-reload` are both false, verify that the code that handles both is not in the bundle.