[WIP] Configure environment vars for start/build commands (#402)
### 🔦 Summary

The Big high-level change that we're introducing here is to be able and distinguish between minification and production transformations. Today the CI runs a slightly different build than the local one. We understand the confusion and we would like to be able to **run the same commands from the CI and Localy** and get the **same result**. It will make a more **predictable** process that is **easier to debug**.

In order to be backward compatible and non breaking for the CI the build will do allmost the same as it does today (that would be similar on CI and Local machines). We'll add a new `build --no-min` which will be faster for debugging and running e2e tests locally.
________________

When running `build` yoshi will create:
* A **production** bundle.
* A **production** **minified** bundle.

When running `build --no-min` yoshi will create:
* A **production** bundle.

When running `start` yoshi will create:
* A **development** bundle.

In the future we could add a `start --min` mode, which will create:
* A **development** **minified** bundle.
______________________

### Changes
* `babel-preset-yoshi`
  * Fix usage of `babel-plugin-transform-react-remove-prop-types`

* `yoshi`
  * Remove leftovers from `caporal`, types and default values for the CLI app that doesn't work anymore with `commander`.
  * Add `--no-min` CLI option.
  * Remove unused files (internal).
  * Run build with `NODE_ENV="production"`.
  * Run test with `NODE_ENV="test"`.
  * Run start with `NODE_ENV="development"`.
  * Configure the Webpack config using `production` & `min` instead of `debug`.
  * Use `localIdentName` from the project itself instead of from `yoshi-runtime`.
  * Remove the preset, and require the commands directly.
  * Change terminology from `webpack-development` to `webpack-debug`.
  * Change the emitted webpack stats file name:
	* `webpack-stats.prod.json` => `webpack-stats.min.json`
	* `webpack-stats.dev.json` => `webpack-stats.json`


### 🗺️ Test Plan
We'll need to adjust the tests to run with the new modes.