validate yoshi config before running command (#535)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
This PR adds configuration validation before running yoshi commands. Fixes #500. This initial implementation is using `ajv` to validate the configuration object against a JSON schema. Then the errors are printed using `YoshiOptionsValidationError` which is a copied implementation from the webpack-cli repository. (The default errors text from `ajv` are not great, so the webpack way seems a good way to go).

Please look carefully at the JSON schema. I created it based on the API docs, but maybe you'll know better. And if there are old deprecated fields, we can print deprecation warning for them, I just didn't know if there are any.

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
So this part is still is open for discussion. What testing strategy do you see for this if at all? I can add unit tests for the copied webpack code for generating the error messages, but maybe it won't add too much, and a full e2e test may be an overkill. Thoughts?