Simplify i18n setup to use webpack's dynamic imports (#542)
### 🔦 Summary

This PR attempts to simplify how `i18next` is configured in a project: Instead of having static files available it uses Webpack's code splitting to fetch resources. 

That way the user can control which translation should go into the bundle and which ones should be fetched in runtime (or even prefetched/preloaded).

It removes the need to copy files from one directory to another (by Yoshi). It also removes the need to configure `staticsBaseUrl` with `i18next` (also for GA'd libraries) as the bundle is already built with the correct publicUrl (since #455).

It also uses i18next's namespaces by default as a way of reducing one large translation file.