Simplify server code and add basic annotations to it (#549)
### 🔦 Summary

This PR simplifies the server code and adds some explanations on things that are not immediately clear.

I think that annotating confusing parts in the generated code can help in making it more approachable. In the end, this code is going to be part of the users' projects. Our users should understand the code that's inside their own projects.

This is something other popular boilerplates/generators do and I think that even small explanations can help.