Generate git repo when requested (#540)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
Fixes #521. In this PR, we `git init` a new generated project if the user answered yes to the question `Is this a git repository?`. I've added the question to support generating a project without a git repository as a very early support for generating a project inside a monorepo. Adding the question relates to issues like #456, allowing us to also not generate certain parts if we don't need a git repo (like not installing husky).

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
Added e2e tests to make sure a git repo has been initialised if requested.