prefer local yoshi installation when using cli (#588)
### 🔦 Summary
This patch will use the `import-local` package, in order to prefer a local `yoshi` installation, over a global one, while running the cli.

If a user is running a `yoshi XXX` command from his terminal, and he has a global `yoshi` installation, he will get his global `yoshi` and not his local `yoshi`. This might cause confusion, in cases where the global and local versions do not match... 

### 🗺️ Test Plan
This is a super tiny (nice to have) feature, which is pretty complicated to test... I did not write a test for it. WDYT? (: