Remove the duplicate package checker plugin (#621)
### 🔦 Summary
As part of the efforts of #512 we've decided that the duplicate package checker produces a lot of clutter when the intent of the user is develop his app and not to audit it.

We can later on put it back under an audit command. When the intent is to audit the app.