Improve e2e tests by emulating production installations (#638)
### 🔦 Summary

This PR changes how our E2E tests install the dependencies of generated projects.

Instead of installing dependencies and then copying/linking the version of Yoshi/others, it creates a local npm registry and publishes the entire monorepo to it. It then installs all dependencies from that local registry to mimic the tree as much as possible.