Fix yoshi not outputing verbose errors on CI (#770)
### 🔦 Summary
This should fix `yoshi` not being verbose when running in CI. It seems like `commander.js` doesn't allow a default value on a bool option as the default value is always false (if the flag doesn't exist). So a quick fix for now is to check for `inTeamCity` when outputing the crash logs.