Pin component libraries templates Storybook version to `alpha.14` for usage with React 15 (#761)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
Fixes #527.
Seems that Storybook uses a newer version of `emotion`, that needs React 16: https://github.com/storybooks/storybook/issues/4191.
Pinning Storybook to a specific version is a quick and dirty fix that can be used until we'll support React 16

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
It shouldn't change much. We only need to check if storybook can actually run on a new yoshi project, using React 15.