Start a forward proxy to intercept public cdn requests to the local cdn (#735)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
Fixes #707. `yoshi` will start a forward proxy that will intercept requests made for the public cdn, and will proxy them to the local cdn instead. This solution fixes the problems both for puppeteer and protractor. 

For protractor - this PR configures the tests to run with chrome and start it with the proxy. If a project is configuring their own browser, they may not get this configuration and will need to do it themselves.

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
Added a test for protractor, that runs a CI build and test of a project with dynamic imports, and makes sure the tests pass. also made sure that without the new changes the new test doesn't pass :)