Ignore ssl errors when proxying requests to the local cdn (#835)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
Fixes an issue with the forward proxy when the local cdn is started with ssl and a self signed certificate. It will make the proxy ignore errors in ssl certificates.

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
Added a test case for protractor with a local cdn that is started with ssl.